pyproject.toml
[project]
name = "mlops-lab1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.0",
    "Pillow>=9.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pylint>=2.15",
    "black>=23.0",
    "uvicorn>=0.38.0",
    "fastapi>=0.119.1",
    "jinja2>=3.1.6",
    "flask>=3.1.2",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "gradio>=5.49.1",
    "requests>=2.32.5",
]
